add process_feat staticmethod on feature dataset
